requirements.txt
Django==2.1.11
djangorestframework==3.10.3
django-filter==2.2.0
Markdown==3.1.1
pytest==5.1.2
pytest-django==3.5.1
pytest-cov==2.7.1
python-decouple==3.1
dj-static==0.0.6
gunicorn==19.9
requests==2.22.0
